requirements.txt
requests
beautifulsoup4
opencv-python
numpy
pillow
pytesseract